LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/websocket.https-expected.txt

CONSOLE MESSAGE: WebSocket network error: OSStatus Error -9807: Invalid certificate chain 
CONSOLE MESSAGE: WebSocket network error: OSStatus Error -9807: Invalid certificate chain 


FAIL Verify WebSocket handshake channel does not get intercepted assert_unreached: unexpected rejection: [object Event] Reached unreachable code
